# docs/collection/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Sistema de Autoatendimento da Lanchonete",
        "description": "Documentacao automatica via Swagger e Redoc",
        "version": "1.0.0"
    },
    "paths": {
...
        "/pedidos/": {
            "get": {
                "tags": ["pedidos"],
                "summary": "Listar Pedidos",
                "operationId": "listar_pedidos_pedidos__get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/PedidoResponseSchema"
                                    },
                                    "type": "array",
                                    "title": "Response Listar Pedidos Pedidos  Get"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "example": {
                                    "message": ""
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": ["pedidos"],
                "summary": "Criar Pedido",
                "operationId": "criar_pedido_pedidos__post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PedidoCreateSchema"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PedidoProdutosResponseSchema"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "example": {
                                    "message": "Erro de integridade ao salvar o pedido | Erro de integridade ao salvar produtos no pedido"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/pedidos/{pedido_id}": {
            "get": {
                "tags": ["pedidos"],
                "summary": "Buscar Pedido",
                "operationId": "buscar_pedido_pedidos__pedido_id__get",
                "parameters": [{
                    "name": "pedido_id",
                    "in": "path",
                    "required": true,
                    "schema": {
                        "type": "integer",
                        "title": "Pedido Id"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PedidoProdutosResponseSchema"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "example": {
                                    "message": "Pedido não encontrado | Produto(s) do pedido não encontrado(s)"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "example": {
                                    "message": ""
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": ["pedidos"],
                "summary": "Atualizar Pedido",
                "operationId": "atualizar_pedido_pedidos__pedido_id__put",
                "parameters": [{
                    "name": "pedido_id",
                    "in": "path",
                    "required": true,
                    "schema": {
                        "type": "integer",
                        "title": "Pedido Id"
                    }
                }],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PedidoAtualizaSchema"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PedidoResponseSchema"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "example": {
                                    "message": "Pedido não encontrado"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "example": {
                                    "message": "Pedido já finalizado"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": ["pedidos"],
                "summary": "Deletar Pedido",
                "operationId": "deletar_pedido_pedidos__pedido_id__delete",
                "parameters": [{
                    "name": "pedido_id",
                    "in": "path",
                    "required": true,
                    "schema": {
                        "type": "integer",
                        "title": "Pedido Id"
                    }
                }],
                "responses": {
                    "204": {
                        "description": "Pedido deletado com sucesso",
                        "content": {
                            "application/json": {
                                "example": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "example": {
                                    "message": "Pedido não encontrado"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "example": {
                                    "message": "Erro de integridade ao deletar o pedido"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/status_pedido/": {
            "get": {
                "tags": ["status_pedido"],
                "summary": "Listar Todos",
                "operationId": "listar_todos_status_pedido__get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/StatusPedidoResponseSchema"
                                    },
                                    "type": "array",
                                    "title": "Response Listar Todos Status Pedido  Get"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "example": {
                                    "message": "Erro de integridade ao buscar os statuses"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": ["status_pedido"],
                "summary": "Criar",
                "operationId": "criar_status_pedido__post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StatusPedidoCreateSchema"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StatusPedidoResponseSchema"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "example": {
                                    "message": "Erro de integridade ao criar o status"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/status_pedido/{id}": {
            "get": {
                "tags": ["status_pedido"],
                "summary": "Buscar Status",
                "operationId": "buscar_status_status_pedido__id__get",
                "parameters": [{
                    "name": "id",
                    "in": "path",
                    "required": true,
                    "schema": {
                        "type": "integer",
                        "title": "Id"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StatusPedidoResponseSchema"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "example": {
                                    "message": "Erro de integridade ao buscar o status"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "example": {
                                    "message": "Status não encontrado"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": ["status_pedido"],
                "summary": "Atualizar",
                "operationId": "atualizar_status_pedido__id__put",
                "parameters": [{
                    "name": "id",
                    "in": "path",
                    "required": true,
                    "schema": {
                        "type": "integer",
                        "title": "Id"
                    }
                }],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StatusPedidoUpdateSchema"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StatusPedidoResponseSchema"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "example": {
                                    "message": "Status não encontrado"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "example": {
                                    "message": "Erro de integridade ao atualizar o status"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": ["status_pedido"],
                "summary": "Deletar",
                "operationId": "deletar_status_pedido__id__delete",
                "parameters": [{
                    "name": "id",
                    "in": "path",
                    "required": true,
                    "schema": {
                        "type": "integer",
                        "title": "Id"
                    }
                }],
                "responses": {
                    "204": {
                        "description": "Status deletado com sucesso",
                        "content": {
                            "application/json": {
                                "example": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "example": {
                                    "message": "Status não encontrado"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "example": {
                                    "message": "Erro de integridade ao excluir o status"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CategoriaProdutoResponseSchema": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "nome": {
                        "type": "string",
                        "title": "Nome"
                    }
                },
                "type": "object",
                "required": ["id", "nome"],
                "title": "CategoriaProdutoResponseSchema"
            },
...
            "ClienteResponseSchema": {
                "properties": {
                    "cliente_id": {
                        "type": "integer",
                        "title": "Cliente Id"
                    },
                    "nome": {
                        "anyOf": [{
                            "type": "string"
                        }, {
                            "type": "null"
                        }],
                        "title": "Nome"
                    },
                    "email": {
                        "anyOf": [{
                            "type": "string"
                        }, {
                            "type": "null"
                        }],
                        "title": "Email"
                    },
                    "telefone": {
                        "anyOf": [{
                            "type": "string"
                        }, {
                            "type": "null"
                        }],
                        "title": "Telefone"
                    },
                    "cpf": {
                        "anyOf": [{
                            "type": "string"
                        }, {
                            "type": "null"
                        }],
                        "title": "Cpf"
                    }
                },
                "type": "object",
                "required": ["cliente_id", "nome", "email", "telefone", "cpf"],
                "title": "ClienteResponseSchema"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "PedidoAtualizaSchema": {
                "properties": {
                    "status": {
                        "type": "integer",
                        "title": "Status"
                    }
                },
                "type": "object",
                "required": ["status"],
                "title": "PedidoAtualizaSchema"
            },
            "PedidoCreateSchema": {
                "properties": {
                    "cliente_id": {
                        "anyOf": [{
                            "type": "integer"
                        }, {
                            "type": "null"
                        }],
                        "title": "Cliente Id"
                    },
                    "produtos": {
                        "items": {},
                        "type": "array",
                        "title": "Produtos"
                    }
                },
                "type": "object",
                "required": ["cliente_id", "produtos"],
                "title": "PedidoCreateSchema"
            },
            "PedidoProdutosResponseSchema": {
                "properties": {
                    "pedido_id": {
                        "type": "integer",
                        "title": "Pedido Id"
                    },
                    "cliente_id": {
                        "$ref": "#/components/schemas/ClienteResponseSchema"
                    },
                    "status": {
                        "$ref": "#/components/schemas/StatusPedidoResponseSchema"
                    },
                    "data_criacao": {
                        "type": "string",
                        "format": "time",
                        "title": "Data Criacao"
                    },
                    "data_alteracao": {
                        "anyOf": [{
                            "type": "string",
                            "format": "time"
                        }, {
                            "type": "null"
                        }],
                        "title": "Data Alteracao"
                    },
                    "data_finalizacao": {
                        "anyOf": [{
                            "type": "string",
                            "format": "time"
                        }, {
                            "type": "null"
                        }],
                        "title": "Data Finalizacao"
                    },
                    "produtos": {
                        "anyOf": [{
                            "items": {
                                "$ref": "#/components/schemas/ProdutoResponseSchema"
                            },
                            "type": "array"
                        }, {
                            "type": "null"
                        }],
                        "title": "Produtos"
                    }
                },
                "type": "object",
                "required": ["pedido_id", "cliente_id", "status", "data_criacao", "data_alteracao", "data_finalizacao", "produtos"],
                "title": "PedidoProdutosResponseSchema"
            },
            "PedidoResponseSchema": {
                "properties": {
                    "pedido_id": {
                        "type": "integer",
                        "title": "Pedido Id"
                    },
                    "cliente_id": {
                        "$ref": "#/components/schemas/ClienteResponseSchema"
                    },
                    "status": {
                        "$ref": "#/components/schemas/StatusPedidoResponseSchema"
                    },
                    "data_criacao": {
                        "type": "string",
                        "format": "time",
                        "title": "Data Criacao"
                    },
                    "data_alteracao": {
                        "anyOf": [{
                            "type": "string",
                            "format": "time"
                        }, {
                            "type": "null"
                        }],
                        "title": "Data Alteracao"
                    },
                    "data_finalizacao": {
                        "anyOf": [{
                            "type": "string",
                            "format": "time"
                        }, {
                            "type": "null"
                        }],
                        "title": "Data Finalizacao"
                    }
                },
                "type": "object",
                "required": ["pedido_id", "cliente_id", "status", "data_criacao", "data_alteracao", "data_finalizacao"],
                "title": "PedidoResponseSchema"
            },
...
            "ProdutoResponseSchema": {
                "properties": {
                    "produto_id": {
                        "type": "integer",
                        "title": "Produto Id"
                    },
                    "nome": {
                        "type": "string",
                        "title": "Nome"
                    },
                    "descricao": {
                        "anyOf": [{
                            "type": "string"
                        }, {
                            "type": "null"
                        }],
                        "title": "Descricao"
                    },
                    "preco": {
                        "type": "string",
                        "title": "Preco"
                    },
                    "categoria": {
                        "$ref": "#/components/schemas/CategoriaProdutoResponseSchema"
                    }
                },
                "type": "object",
                "required": ["produto_id", "nome", "preco", "categoria"],
                "title": "ProdutoResponseSchema"
            },
...
            "StatusPedidoCreateSchema": {
                "properties": {
                    "descricao": {
                        "type": "string",
                        "title": "Descricao"
                    }
                },
                "type": "object",
                "required": ["descricao"],
                "title": "StatusPedidoCreateSchema"
            },
            "StatusPedidoResponseSchema": {
                "properties": {
                    "descricao": {
                        "type": "string",
                        "title": "Descricao"
                    },
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    }
                },
                "type": "object",
                "required": ["descricao", "id"],
                "title": "StatusPedidoResponseSchema"
            },
            "StatusPedidoUpdateSchema": {
                "properties": {
                    "descricao": {
                        "type": "string",
                        "title": "Descricao"
                    }
                },
                "type": "object",
                "required": ["descricao"],
                "title": "StatusPedidoUpdateSchema"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [{
                                "type": "string"
                            }, {
                                "type": "integer"
                            }]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": ["loc", "msg", "type"],
                "title": "ValidationError"
            }
        }
    }
}